separate structs to separate files
