+++ src/xcode/ENA/ENA/Source/Scenes/Home/Cells/Statistics/Model/LocalStatisticsRegion.swift
@@ -10,16 +10,3 @@ struct LocalStatisticsRegion: Codable, Equatable {
 	let id: String
 	let regionType: RegionType
 }
-
-enum RegionType: String, CaseIterable, Codable {
-	case federalState
-	case administrativeUnit
-}
-
-
-struct SevenDayData {
-	var regionName: String
-	var id: Int = 0
-	var updatedAt: Int64 = 0
-	var sevenDayIncidence: SAP_Internal_Stats_SevenDayIncidenceData
-}
