+++ src/xcode/ENA/ENA/Source/Scenes/Home/Cells/Statistics/Model/RegionType.swift
@@ -1,0 +1,10 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+
+enum RegionType: String, CaseIterable, Codable {
+	case federalState
+	case administrativeUnit
+}
