+++ src/xcode/ENA/ENA/Source/Scenes/Home/Cells/Statistics/Model/SevenDayData.swift
@@ -1,0 +1,12 @@
+////
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+
+struct SevenDayData {
+	var regionName: String
+	var id: Int = 0
+	var updatedAt: Int64 = 0
+	var sevenDayIncidence: SAP_Internal_Stats_SevenDayIncidenceData
+}
